Refactor and remove a load of unused stuff
This removes the show type command, the insert type command, the HaRe
commands and the case split command – none of them have existed in HIE
for a while now, and none of them are in ghcide/hls either
