+++ src/commands/constants.ts
@@ -1,6 +1,4 @@
 export namespace CommandNames {
-  export const InsertTypeCommandName = 'hie.commands.insertType';
-  export const ShowTypeCommandName = 'hie.commands.showType';
   export const ImportIdentifierCommandName = 'hie.commands.importIdentifier';
   export const RestartHieCommandName = 'hie.commands.restartHie';
 }
