+++ tslint.json
@@ -1,5 +1,4 @@
 {
-  "defaultSeverity": "error",
   "extends": ["tslint:recommended"],
   "jsRules": {},
   "rules": {
